refactor(registar): обједини Konvertor на utils, уклони convert_utils и празан resources (#300)
Постојале су две класе `Konvertor`: груба у `management/commands/convert_utils.py`
(штампа упозорења, без обраде `None`, без докстрингова) и дотерана у
`utils/migration_converters.py` (логовање, обрада `None`/`TypeError`,
докстрингови, плюс `date`/`split_name`). У продукцији је била коришћена
лошија копија, а дотерана само у сопственом тесту.

Уз то, `convert_utils.py` је стајао у `commands/` иако нема `Command`
класу (Django би пукао при покушају извршавања), а пакет `registar/resources/`
је био празан и неискоришћен.

- 3 увоза (`migracija_svestenika`, `migracija_ulice_svestenik`,
  `migracija/helpers`) препоинтирана на `registar.utils.migration_converters`
  (исти `string`/`int` API, само бољи)
- обрисан `management/commands/convert_utils.py`
- обрисан празан пакет `registar/resources/`
