+++ crkva/registar/management/commands/migracija_ulice_svestenik.py
@@ -17,8 +17,8 @@ import zipfile
 
 from django.core.management.base import BaseCommand
 from django_tenants.utils import schema_context
-from registar.management.commands.convert_utils import Konvertor
 from registar.models import Adresa, Svestenik
+from registar.utils.migration_converters import Konvertor
 
 
 def _read_dbf(raw: bytes):
